READY FOR REVIEW (Functions_good)
Code was made to include fahrenheit as default. However existing if blocks still input a value into the convertion function. Is it possible to give None type as a value to the function so it will just run the default target_scale, or do we have to use an if block to compare if there is a correct or incorrect input?